Fix data leakage issue in previous final version
During the deployment and testing phase, we found that the model trained as described in the previous paper did not achieve the expected accuracy in real-world applications. After carefully inspecting the dataset, we discovered a data leakage issue. We then re-collected the dataset and improved our experiments on data augmentation strategies based on the original setup. Additionally, we introduced more augmentation combinations to enable a more comprehensive comparison. After addressing these issues, we obtained results that are more intuitive and convincing.